Refactor getUpdatedHubSync()
This method used to always return an oparl specification hub sync
repository. This however is not feasable anymore in a world where
this method should be able to return any repository, updated.

It may in the future also be advisable to move this into a
special HubSync\Repository constructor that handles the updating.