First draft for liboparl building
THIS IS NOT PRODUCTION READY!

Liboparl will be built with the prefix configured in LIBOPARL_PREFIX,
and currently `ninja install` is being run directly from the job
runner which would require it to gain much more capabilities than it
should have.

In the future, this will be reworked to mark the build as installable and
use an extra, completely separate cronjob/script which will do the actual
installation.