Basic endpoint validation integration
This commit finishes the pre-validation actions for OParl endpoint
validation. The request form and messages are now formatted and
internationalized (localized for German) correctly. Additionally,
the validation job is already issueing the correct validator
call, ready to process it's output.

Further work on this will be necessary once the validator is
actually outputting data in the expected format.