Added a levenshtein-based fuzzy route matcher
This is more of a concept than anything. Some routes are close
to other routes that the user may be trying to reach but can't
remmeber the correct uri. The fuzzy matcher tries to guess the
correct route and redirects to there.

This should probably be expanded into a custom 404 handler which
displays the option instead of blindly redirecting the user and
not telling them they were wrong.