Fix encoding on schema files
We write the Schema files as UTF-8 encoded JSON. The controller
previously delivered those with ASCII encoding which resulted
in a few encoding errors in the display.

As a side note: This commit also enables cross origin requests on
the schema 'api' as that might be useful for external tooling.