Improve endpoint update failure handling
Updating endpoint information is by nature prone to
fail. While this commit on the surface appears to
remove try-catch blocks and thus increase the potential
error surface, those could be removed by switching
from Validator::validate() - which throws on
errors - to Validator::fails(). This change also enables
logging the actual validation errors instead
of a not actually helpful exception.

Additional changes include some resilience against
missing name fields in a body (which is a spec
validation but needs to be handled non-the-less) and
bailing out of an update faster if a step fails. The
latter especially applies to not even trying to fetch
a OParl:Body if the OParl:System fetch already failed.