Prepare endpoints UI refactoring
The endpoint listing as of now does quite a lot of not actually
useful stuff. To actually find a OParl:Body, multiple clicks are
required. There is no indication of where (in Germany) those bodies
are and there might even be other things which can enhance the
user experience.

This commit merely prepares the API to support querying the bodies
of a single known endpoint instead of always returning all of them.
Additionally, the endpoint action was moved to a separate controller
to be ready for more endpoint related views if they come.